% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets.R
\name{Chaff}
\alias{Chaff}
\title{Annual laying date of breeding common chaffinch (Fringilla coelebs).}
\format{
A data frame with 47 rows and 3 variables.
\describe{
  \item{Year}{Year of laying date measurement.}
  \item{Date}{Average date of measurement.}
  \item{Laydate}{Average annual laying date in days after January 1st.}
  }
}
\description{
Average annual laying date of common chaffinch (Fringilla coelebs)
measured over 47 years.
}
