% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/daily_data.R
\name{daily_data}
\alias{daily_data}
\title{Convert climate data to a daily format.}
\usage{
daily_data(date, climate, scale)
}
\arguments{
\item{date}{The climate date variable (dd/mm/yyyy). Please specify the parent
environment and variable name (e.g. Climate$Date).}

\item{climate}{The climate variables of interest. 
Please specify the parent environment and variable name (e.g. Climate$Temp).}

\item{scale}{The temporal scale of the original dataset. Either "week" or "month".}
}
\value{
Returns a data frame with daily date information and corresponding climate data.
Where a single climate record is available per week/month, this value will be replicated
across all other days in that same week/month.
}
\description{
Convert data from a monthly or weekly format into a daily format for use in climwin.
}
\examples{
\dontrun{

#Load data
data(Monthly_data)

#View original data to see monthly format.
head(Monthly_data)

#Convert data into a daily data format.
newdat <- daily_data(date = Monthly_data$Date, climate = Monthly_data$Temp, scale = "month")

#View new data to see daily format.
head(newdat, 35)

#Note that data from each month has been duplicated across all days within the month.
#This is needed to work within climwin. Results will not be impacted.

}

}
\author{
Liam D. Bailey and Martijn van de Pol
}
