% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{Mass}
\alias{Mass}
\title{Blue tit body mass}
\format{
A data frame with 47 rows and 3 columns:
\describe{
  \item{Date}{Measurement date, character string in DD/MM/YYYY format.}
  \item{Mass}{Nestling body mass (g).}
  \item{Age}{Nestling age (days) at measurement.}
}
}
\usage{
data(Mass)
}
\description{
Annual body mass measurements of blue tit (\emph{Cyanistes caeruleus})
nestlings, used as the biological response variable in package examples.
}
\seealso{
[MassClimate] for the corresponding climate data.
}
\keyword{datasets}
