% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/datasets.R
\name{MassRand}
\alias{MassRand}
\title{Example output dataframe from function randwin.}
\format{A data frame with 10,302 rows and 14 variables.
\describe{
  \item{ModelAICc}{AICc of fitted climate windows.}
  \item{baselineAICc}{AICc of model without climate included.}
  \item{WindowOpen}{The start day of each tested climate window. Furthest from the biological record.}
  \item{WindowClose}{The end day of each tested climate window. Closest to the biological record.}
  \item{ModelBeta}{Beta estimate of the relationship between temperature and chick mass.}
  \item{ModelBetaQ}{Quadratic beta estimate of the relationship between temperature and chick mass.}
  \item{ModelBetaC}{Cubic beta estimate of the relationship between temperature and chick mass.}
  \item{ModelInt}{Model intercept.}
  \item{furthest}{Furthest day back considered in climatewin.}
  \item{closest}{Closest day back considered in climatewin.}
  \item{Statistics}{The statistic used to analyse climate (e.g. mean, max, slope).}
  \item{Function}{The function used to fit climate (e.g. linear ("lin"), quadratic ("quad"))}
  \item{FixedWindow}{TRUE or FALSE, whether a fixed window was tested.}
  \item{Cutoff_day,Cutoff_month}{If FixedWindow is TRUE, the date from which the climate window was tested.}
  \item{Randomised}{Whether the data was generated by \code{\link{climatewin}} or \code{\link{randwin}}.}
  \item{Repeat}{The number of randomisations carried out.}
  }}
\description{
Output file from function \code{\link{randwin}} using temperature and chick mass data.
Generated with \code{\link{Mass}} and \code{\link{MassClimate}} dataframes.
}

