% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets.R
\name{Monthly_data}
\alias{Monthly_data}
\title{Example of data in a monthly format.}
\format{A data frame with 576 rows and 2 variables
\describe{
  \item{Date}{Date of measurements (dd/mm/yyyy).}
  \item{Temp}{Temperature data.}
  }}
\description{
Randomly generated data stored in a monthly format.
Used as an example of function 'daily_data'.
}
