% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/plotbest.R
\name{plotbest}
\alias{plotbest}
\title{Visualise the best climate window}
\usage{
plotbest(Dataset, BestModel, BestModelData)
}
\arguments{
\item{Dataset}{A dataframe containing information on all fitted climate
windows. Output from \code{\link{climatewin}}.}

\item{BestModel}{A model object. The strongest climate window model. Output
from \code{\link{singlewin}} or \code{\link{climatewin}}.}

\item{BestModelData}{A dataframe with the data used to
fit the strongest climate window model. Output from \code{\link{singlewin}}
or \code{\link{climatewin}}.}
}
\value{
Returns a scatterplot with a fitted line to show the fit of the best
 model through the data.
}
\description{
Create a scatterplot showing the fit of the best climate window model through
the biological data.
}
\examples{
# Visualise the best climate window from the datasets Mass and MassClimate

data(MassOutput)
data(Mass)
data(MassClimate)

single <- singlewin(Xvar = MassClimate$Temp, Cdate = MassClimate$Date, Bdate = Mass$Date,
                   baseline = lm(Mass ~ 1, data = Mass),furthest = 72, closest = 15,
                   stat = "mean", func = "lin",
                   type = "fixed", cutoff.day = 20, cutoff.month = 5,
                   Cmissing = FALSE, Cinterval = "day")

plotbest(Dataset = MassOutput, BestModel = single$BestModel,
        BestModelData = single$BestModelData)
}
\author{
Liam D. Bailey and Martijn van de Pol
}

