% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pvalue.R
\name{pvalue}
\alias{pvalue}
\title{Determine the probability that a given climate signal is 'true'}
\usage{
pvalue(rand.dataset, full.dataset, metric, sample.size)
}
\arguments{
\item{rand.dataset}{Output dataframe of function \code{\link{randwin}}.}

\item{full.dataset}{Output dataframe of function \code{\link{climatewin}}.}

\item{metric}{"AIC" or "Weight". Determine whether a value of PDAICc or
Pw will be returned.}

\item{sample.size}{Sample size of analysis. Relevant for metric Pw only.}
}
\value{
Returns a value representing the probability that a given climate 
 window result is a false positive.#'
}
\description{
Calculate either PDAICc or Pw of a given climate window. See ___
for more detail
}
\author{
Liam D. Bailey and Martijn van de Pol
}

