% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/climwin_class.R
\name{climwin_weightwin}
\alias{climwin_weightwin}
\title{climwin_weightwin S7 object output from run_weightwin}
\usage{
climwin_weightwin(
  weightwin_summary = (function (.data = list(), row.names = NULL) 
 {
     if
    (is.null(row.names)) {
         list2DF(.data)
     }
     else {
         out <-
    list2DF(.data, length(row.names))
         attr(out, "row.names") <- row.names
      
      out
     }
 })(),
  weightwin_output = list(),
  range = integer(0)
)
}
\description{
Allows us to have generic print and plot methods.
}
\examples{
## Create a climwin object from `run_weightwin`
Climate <- read.csv(system.file("MassClimate.csv", package = "climwin"))
Mass <- read.csv(system.file("Mass.csv", package = "climwin"))
results <- run_slidingwin(range = 0:2, 
                        climate_data = Climate, 
                        bio_data = Mass,
                        basemodel = lm(Mass ~ climate, data = bio_data))

## Same method names exist, but behave differently internally
getDataset(results)
getBestModel(results)

## Use generic methods
## Using 'plot' is the same as 'plot_slidingwin'
plot(results, y = "Mass")

}
