% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/date_conversion.R
\name{convert_dates_to_int}
\alias{convert_dates_to_int}
\title{Convert Between Date Formats}
\usage{
convert_dates_to_int(dates, min_date = NULL)
}
\arguments{
\item{dates}{A vector of dates in character format (DD/MM/YYYY)}

\item{min_date}{Optional. The reference date to use as day 0. If NULL, uses the earliest date in the input.}
}
\value{
A list containing:
  \item{date_int}{Integer vector of days since min_date}
  \item{min_date}{The reference date used (as Date object)}
  \item{lookup_table}{Data frame mapping integers to character dates}
}
\description{
This function converts dates between character strings (DD/MM/YYYY) and integers (days since earliest date).
It also provides a lookup table for converting back to character format.
}
\examples{
dates <- c("01/01/1979", "02/01/1979", "03/01/1979")
result <- convert_dates_to_int(dates)

}
