% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit_climate_models.R
\name{fit_climate_models}
\alias{fit_climate_models}
\title{Fit Linear Models Between Mass and Climate Data}
\usage{
fit_climate_models(
  climate_means,
  bio_data,
  basemodel,
  mass_col = "Mass",
  date_col = "Date"
)
}
\arguments{
\item{climate_means}{A LIST of data frames, each containing Bio_Date, Start_Date, End_Date, and Summary_Value columns}

\item{bio_data}{A data frame containing biological data with a Date column. bio_data must also contain a 'climate' column (can be initialized with zeros).}

\item{basemodel}{An lm model object that will be updated for each climate window (e.g., lm(Mass ~ climate, data = bio_data)). This argument is required and cannot be NULL.}

\item{mass_col}{Character string specifying the name of the mass column in bio_data}

\item{date_col}{Character string specifying the name of the date column in bio_data}
}
\value{
A data frame containing:
        - Start_Date: Start date of the climate window
        - End_Date: End date of the climate window
        - AIC: AIC value for the linear model
        - R_squared: R-squared value for the model
        - Slope: Slope of the relationship
        - P_value: P-value for the slope
}
\description{
This function takes a list of data frames (output from splitting calculate_temp_means)
and fits linear models between mass and climate data for each combination of days,
returning AIC values and other statistics. The function requires a base model structure
that will be updated for each climate window.
}
\examples{
# Example usage:
climate_means <- calculate_temp_means(0:2, bio_data = Mass)
Mass$climate <- 0
basemodel <- lm(Mass ~ climate, data = Mass)
results <- fit_climate_models(climate_means, Mass, basemodel = basemodel)

}
