% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit_climate_models.R
\name{fit_climate_models}
\alias{fit_climate_models}
\title{Fit Linear Models Between Mass and Climate Data}
\usage{
fit_climate_models(climate_means, basemodel, bio_data, parallel = FALSE)
}
\arguments{
\item{climate_means}{A LIST of data frames, each containing Bio_Date, Start_Date, End_Date, and Summary_Value columns}

\item{basemodel}{An lm model object that will be updated for each climate window (e.g., lm(Mass ~ climate, data = bio_data)). This argument is required and cannot be NULL.}

\item{bio_data}{A data frame containing all variables used in the basemodel. This argument is required and cannot be NULL.}

\item{parallel}{Logical. If TRUE, parallel processing is used. Default is TRUE.}
}
\value{
A data frame containing:
        - Start_Date: Start date of the climate window
        - End_Date: End date of the climate window
        - Start_Day: Start day as integer (number of days before Bio_Date)
        - End_Day: End day as integer (number of days before Bio_Date)
        - AIC: AIC value for the linear model
}
\description{
This function takes a list of data frames (output from splitting calculate_temp_means)
and fits linear models between mass and climate data for each combination of days,
returning AIC values and other statistics. The function requires a base model structure
that will be updated for each climate window.
}
\examples{
# Example usage:
Climate <- read.csv("MassClimate.csv")
Mass <- read.csv("Mass.csv")
climate_means <- calculate_temp_means(0:2, climate_data = Climate, bio_data = Mass)
results <- fit_climate_models(climate_means, basemodel = lm(Mass ~ climate, data = bio_data), bio_data = Mass)

}
