% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/datasets.R
\name{Mass}
\alias{Mass}
\title{Body mass data since 1979.}
\format{A data frame with 47 rows and 2 variables
\describe{
  \item{Date}{Date of mass measurements (dd/mm/yyyy).}
  \item{Mass}{Annual average body mass in grams.}
  }}
\description{
Artificially generated data representing
average body mass since 1979.
}

