% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets.R
\name{Mass}
\alias{Mass}
\title{Chick body mass data since 1979.}
\format{
A data frame with 47 rows and 2 variables
\describe{
  \item{Date}{Date of mass measurements (dd/mm/yyyy).}
  \item{Mass}{Annual average body mass in grams.}
  \item{Age}{Annual average age of mother in years.}
  }
}
\description{
Artificially generated data representing
average body mass of bird chicks since 1979.
}
