% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets.R
\name{MassRand}
\alias{MassRand}
\title{Example output dataframe from function randwin.}
\description{
Output file from function \code{\link{randwin}} using temperature and mass data. 
Generated with \code{\link{Mass}} and \code{\link{MassClimate}} dataframes.
}

