% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets.R
\name{MassRand}
\alias{MassRand}
\title{Example output dataframe from function randwin.}
\format{
A data frame with 5 rows and 21 variables.
\describe{
  \item{deltaAICc}{Difference between model AICc of fitted climate window and a null model containing no climate.}
  \item{WindowOpen}{The start day of each tested climate window. Furthest from the biological record.}
  \item{WindowClose}{The end day of each tested climate window. Closest to the biological record.}
  \item{ModelBeta}{Beta estimate of the relationship between temperature and mass.}
  \item{Std.Error}{Standard error term for linear model betas.}
  \item{ModelBetaQ}{Quadratic beta estimate of the relationship between temperature and mass.}
  \item{ModelBetaC}{Cubic beta estimate of the relationship between temperature and mass.}
  \item{ModelInt}{Model intercept.}     
  \item{Function}{The function used to fit climate (e.g. linear ("lin"), quadratic ("quad"))}
  \item{Furthest}{Furthest day back considered in slidingwin.}
  \item{Closest}{Closest day back considered in slidingwin.}
  \item{Statistics}{The aggregate statistic used to analyse climate (e.g. mean, max, slope).}
  \item{Type}{Whether "fixed" or "variable" climate windows were tested.}
  \item{K}{Number of folds used for k-fold cross validation.}
  \item{ModWeight}{Model weight of each fitted climate window.}
  \item{sample.size}{Sample size (i.e. number of years or sites) used for climate window analysis.}
  \item{Reference.day,Reference.month}{If type is "absolute", the date from which the climate window was tested.}
  \item{Randomised}{Whether the data was generated using \code{\link{slidingwin}} or \code{\link{randwin}}.}
  \item{Repeat}{The number of randomisations carried out.}
  \item{WeightDist}{Model spread of 95 percent confidence set of models.}
  }
}
\description{
Output file from function \code{\link{randwin}} using temperature and mass data. 
Generated with \code{\link{Mass}} and \code{\link{MassClimate}} dataframes.
}
