% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/medwin.R
\name{medwin}
\alias{medwin}
\title{Return the value of the median climate window}
\usage{
medwin(Dataset, CW = 0.95)
}
\arguments{
\item{Dataset}{Output from function \code{\link{climatewin}} with information on all fitted climate windows.}

\item{CW}{The model weight cut-off that determines which climate windows are used to calculate the median values.}
}
\value{
Returns a list item containing median values of window open and window close
}
\description{
Returns the median values for window open and window close
from all climate windows that fall within specified model weights.
}
\examples{
Find median window values for climate windows fitted to the Mass dataset

data(MassOutput)

medwin(Dataset = MassOutput, CW = 0.95)
}
\author{
Liam D. Bailey and Martijn van de Pol
}

