% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/medwin.R
\name{medwin}
\alias{medwin}
\title{Determine the median start and end time for climate windows}
\usage{
medwin(dataset, cw = 0.95)
}
\arguments{
\item{dataset}{Output dataframe of function \code{\link{slidingwin}}.}

\item{cw}{Cut-off for confidence set (0.95 by default)}
}
\value{
Returns two values representing the median start and end
time of climate windows within the confidence set.
}
\description{
Determine the median start and end time for climate windows within
a chosen confidence set.
}
\examples{
# Determine median start and end time of MassOutput from the 95\% confidence set

medwin(MassOutput, cw = 0.95)

}
\author{
Liam D. Bailey and Martijn van de Pol
}
