% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets.R
\name{Monthly_data}
\alias{Monthly_data}
\title{Monthly temperature data}
\format{
A data frame with 576 rows and 2 variables
\describe{
  \item{Date}{Date of temperature measurements (dd/mm/yyyy).}
  \item{Temp}{Mean monthly temperature}
  }
}
\description{
Artificially generated temperature data
at a monthly scale. Used for code testing.
}
