% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets.R
\name{Offspring}
\alias{Offspring}
\title{Reproductive success of birds since 2009.}
\format{
A data frame with 1,619 rows and 5 variables.
\describe{
  \item{Offspring}{Total number of offspring produced.}
  \item{Date}{Date of hatching (dd/mm/yyyy).}
  \item{Order}{Order of nest within each season.}
  \item{BirdID}{Individual ID of female.}
  \item{Cohort}{Grouping factor designating the breeding season
                 of each record.}
   }
}
\description{
Artificially generated data representing
reproductive success of birds since 2009.
}
