% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets.R
\name{OffspringClimate}
\alias{OffspringClimate}
\title{Daily climate data since 2009.}
\format{
A data frame with 2,588 rows and 3 variables.
\describe{
  \item{Date}{Date when climate was recorded (dd/mm/yyyy).}
  \item{Rain}{Daily rainfall data in mm.}
  \item{Temperature}{Daily temperature data in degrees centigrade.}
  }
}
\description{
Daily temperature and rainfall data since 2009.
Coincides with biological data from \code{\link{Offspring}}.
}
