% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_best.R
\name{plot_best}
\alias{plot_best}
\title{Plot Best Model Fit}
\usage{
plot_best(dataset, x, y, ...)
}
\arguments{
\item{dataset}{Output from run_slidingwin (a list with 'dataset' and 'bestModel' items)}

\item{x}{Character. Name of x variable to plot. If not provided, will be "climate".}

\item{...}{Additional arguments passed to 'predict' function to create model prediction line.}
}
\value{
A ggplot object showing the scatter plot with fitted line
}
\description{
This function creates a scatterplot showing the relationship between the climate variable
and biological response from the best fitting model (lowest AIC) identified by run_slidingwin.
The plot includes the data points and the fitted regression line.
}
\examples{
# Example usage:
Climate <- read.csv(system.file("MassClimate.csv", package = "climwin"))
Mass <- read.csv(system.file("Mass.csv", package = "climwin"))
results <- run_slidingwin(range = 0:100, 
                        climate_data = Climate, 
                        bio_data = Mass,
                        basemodel = lm(Mass ~ climate, data = bio_data))
plot_best(results)

}
