% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_weibull.R
\name{plot_weibull}
\alias{plot_weibull}
\title{Plot Weibull Distribution Weights}
\usage{
plot_weibull(dataset)
}
\arguments{
\item{dataset}{Output from run_weightwin containing weights information in the @weights slot}
}
\value{
A ggplot object showing the Weibull distribution weights
}
\description{
This function creates a line plot visualization of the Weibull distribution weights
used to create the weighted climate variable in the best model. The plot shows how
the weights are distributed across the time range, with Day on the x-axis and 
weight values on the y-axis.
}
\examples{
# Example usage:
data("MassClimate")
data("Mass")
results <- run_weightwin(range = 0:100, 
                        bio_data = Mass,
                        climate_data = MassClimate,
                        cdate = "Date", bdate = "Date",
                        xvar = "Temp",
                        basemodel = lm(Mass ~ climate, data = bio_data),
                        par = c(1.25, 0.5))
plot_weibull(results)

}
