% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotdelta.R
\name{plotdelta}
\alias{plotdelta}
\title{Plot deltaAICc or other measure of model fit}
\usage{
plotdelta(
  dataset,
  fitStat = "deltaAICc",
  arrow = FALSE,
  plotall = FALSE,
  plotallenv,
  ThreeD = FALSE
)
}
\arguments{
\item{dataset}{A dataframe containing information on all fitted climate 
windows. Output from \code{\link{slidingwin}}.}

\item{fitStat}{A character string giving name of measure model fit to plot 
from dataset of fitted climate windows.}

\item{arrow}{TRUE or FALSE. Add arrows to plots to pinpoint best window.}

\item{plotall}{Used in conjunction with function \code{\link{plotall}}. 
Should not be changed manually.}

\item{plotallenv}{Used in conjunction with function \code{\link{plotall}}.
Should not be changed manually.}

\item{ThreeD}{TRUE or FALSE. Generate a 3-dimensional plot of the deltaAICc landscape.}
}
\value{
Returns a colour plot of model deltaAICc values (larger negative
values indicate stronger models). DeltaAICc is the difference between AICc
of each climate window and a null model.
}
\description{
Create a colour plot of model deltaAICc values.
}
\examples{
# Plot deltaAICc estimates for climate windows in the Mass dataset

data(MassOutput)

plotdelta(dataset = MassOutput)
}
\author{
Liam D. Bailey and Martijn van de Pol
}
