% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/plotdelta.R
\name{plotdelta}
\alias{plotdelta}
\title{Plot deltaAICc of models}
\usage{
plotdelta(Dataset)
}
\arguments{
\item{Dataset}{A dataframe containing information on all fitted climate
windows. Output from \code{\link{climatewin}}.}
}
\value{
Returns a colour plot of model deltaAICc values (larger negative
  values indicate stronger models). DeltaAICc is the difference between AICc
  of each climate window and a null model.
}
\description{
Create a colour plot of model deltaAICc values.
}
\examples{
# Plot deltaAICc estimates for climate windows in the Mass dataset

data(MassOutput)

plotdelta(Dataset = MassOutput)
}
\author{
Liam D. Bailey and Martijn van de Pol
}

