% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_slidingwin.R
\name{run_slidingwin}
\alias{run_slidingwin}
\title{Run Sliding Window Analysis Between Mass and Climate Data}
\usage{
run_slidingwin(
  range,
  climate_data,
  bio_data,
  basemodel,
  cdate = "Date",
  bdate = "Date",
  xvar = "Temp",
  fn = mean,
  type = "relative",
  refday = NULL,
  spatial = NULL,
  parallel = FALSE,
  progress = TRUE,
  .basemodelIsCall = FALSE
)
}
\arguments{
\item{range}{A numeric vector specifying the number of days to look back from each date in bio_data.
For example, 0 represents the date itself, while 100 represents 100 days before that date.}

\item{climate_data}{A data frame containing climate data. Required.}

\item{bio_data}{A data frame containing biological data with a date column. Required.}

\item{basemodel}{An lm model object that will be updated for each climate window (e.g., lm(Mass ~ climate, data = bio_data)). Required.}

\item{cdate}{Character string specifying the name of the date column in climate_data. Defaults to "Date".}

\item{bdate}{Character string specifying the name of the date column in bio_data. Defaults to "Date".}

\item{xvar}{Character string specifying the name of the climate variable column in climate_data. Defaults to "Temp".}

\item{fn}{A function to use for summarizing the climate data. Defaults to mean().}

\item{type}{Character string specifying the type of date range calculation. Must be either "relative" (default) or "absolute".}

\item{refday}{Character string in format "DD/MM/YYYY" specifying the reference date to use when type is "absolute".}

\item{parallel}{Logical. If TRUE, parallel processing is used. Default is FALSE.}
}
\value{
A data frame containing:
        - Start_Date: Start date of the climate window
        - End_Date: End date of the climate window
        - Start_Day: Start day as integer (number of days before Bio_Date)
        - End_Day: End day as integer (number of days before Bio_Date)
        - AIC: AIC value for the linear model
}
\description{
This function fits linear models between mass and climate data for each combination of days,
returning AIC values and other statistics. The function requires a base model structure
that will be updated for each climate window. It internally calls calc_windows
to compute climate summaries for each window.
}
\examples{
# Example usage:
Climate <- read.csv(system.file("MassClimate.csv", package = "climwin"))
Mass <- read.csv(system.file("Mass.csv", package = "climwin"))
results <- run_slidingwin(range = 0:2, 
                        climate_data = Climate, 
                        bio_data = Mass,
                        basemodel = lm(Mass ~ climate, data = bio_data))
                        
Mass$site <- sample(c("A", "B"), size = nrow(Mass), replace = TRUE)
Climate1 <- Climate
Climate1$site <- "A"
Climate2 <- Climate
Climate2$site <- "B"
Climate_site <- rbind(Climate1, Climate2)
results_spatial <- run_slidingwin(range = 0:2, 
                        climate_data = Climate_site, 
                        bio_data = Mass,
                        basemodel = lm(Mass ~ climate, data = bio_data),
                        spatial = "site")

}
